@@ -1,9 +1,8 @@
 package com.tisen.smile.activity;
 
-import android.app.Activity;
-import android.support.annotation.LayoutRes;
-import android.support.v7.app.AppCompatActivity;
 import android.view.WindowManager;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * Created by tisen on 2016/9/29.
@@ -12,12 +11,13 @@ public class BaseActivity extends AppCompatActivity {
     public static String TAG = "BaseActivity";
 
     @Override
-    public void setContentView(@LayoutRes int layoutResID) {
+    public void setContentView(int layoutResID) {
         super.setContentView(layoutResID);
         getWindow().addFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS);
         getWindow().addFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_NAVIGATION);
     }
-    public String getTAG(){
+
+    public String getTAG() {
         return getClass().getSimpleName();
     }
 }
