@@ -3,14 +3,14 @@ package com.tisen.smile.fragment;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.annotation.Nullable;
-import android.support.v4.widget.SwipeRefreshLayout;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
 import android.widget.ListView;
 import android.widget.TextView;
+
+import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import com.tisen.smile.R;
 import com.tisen.smile.activity.ShowJokeActivity;
@@ -22,7 +22,6 @@ import com.tisen.smile.utils.Classification;
  */
 public class SelectedFragment extends SwipeRefreshFragment {
 
-    @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         view = inflater.inflate(R.layout.viewpager_view,null);
